requirements.txt
flask
numpy
pillow
opencv-python
Werkzeug
tensorflow
jupyter
pandas

